+++ src/main/resources/templates/kakao-login.html
@@ -6,8 +6,10 @@
 </head>
 <body>
   <h1>카카오 로그인</h1>
-  <a href="/kakao/login/redirect">
-    <img src="/kakao_login_medium_narrow.png"  alt="kakaoLogin">
-  </a>
+  <form action="/api/members/kakao" method="post">
+    <button type="submit">
+      <img src="/kakao_login_medium_narrow.png" alt="kakaoLogin">
+    </button>
+  </form>
 </body>
 </html>
